添加controller

+++ imooc-ad-service/ad-sponsor/src/main/java/com/imooc/ad/dao/AdUnitRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface AdUnitRepository extends JpaRepository<AdUnit, Long> {
 
-    AdUnit findByPlanIdAndUnitName(Integer planId, String unitName);
+    AdUnit findByPlanIdAndUnitName(Long planId, String unitName);
 
 
 }
